Factorize visual debugger, fix object creation, fix hitboxes update.
