+++ GDJS/Runtime/pixi-renderers/RuntimeInstancesContainerPixiRenderer.ts
@@ -1,0 +1,11 @@
+/*
+ * GDevelop JS Platform
+ * Copyright 2013-2016 Florian Rival (Florian.Rival@gmail.com). All rights reserved.
+ * This project is released under the MIT License.
+ */
+namespace gdjs {
+  export interface RuntimeInstancesContainerPixiRenderer {
+    setLayerIndex(layer: gdjs.Layer, index: float): void;
+    getRendererObject();
+  }
+}
